Organize monad-ethcall crate
